requirements.txt
# Web scraping and parsing
selenium==4.34.2
webdriver-manager==4.0.2
beautifulsoup4==4.13.4
httpx==0.28.1

# Notion integration
notion-client==2.5.0

# Environment and configuration
python-dotenv==1.1.1

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10

# Logging and monitoring
loguru==0.7.3
sentry-sdk==2.40.0

# Web interface
streamlit==1.38.0
pandas==2.2.3
pydeck==0.9.1
folium==0.17.0
streamlit-folium==0.23.0
plotly==5.17.0

# Scientific computing (required by pandas, streamlit)
numpy>=1.26.0,<2.3
scipy>=1.14.0

# Task scheduling
apscheduler==3.10.4

# CLI interface
click==8.1.7

# Development tools
ruff==0.7.4
